requirements.txt
aiohttp==3.7.4
aiosignal==1.2.0
async-timeout==3.0.1
attrs==22.1.0
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.0
configparser==5.3.0
discord==1.7.3
discord.py==1.7.3
feedparser==6.0.10
frozenlist==1.3.1
idna==3.3
multidict==6.0.2
pyaes==1.6.1
pyasn1==0.4.8
requests==2.28.1
rsa==4.9
sgmllib3k==1.0.0
Telethon==1.24.0
typing-extensions==4.3.0
urllib3==1.26.11
yarl==1.8.1
python-dateutil==2.8.2
python-dotenv
